my-azure-function/requirements.txt
azure-functions
azure-eventhub
pg8000